+++ crates/ai/src/pipeline/encode/image/embeddings.rs
@@ -20,8 +20,7 @@ pub async fn image_embedding(paths: Vec<PathBuf>) -> anyhow::Result<Option<Strin
         }
     };
 
-    let db_location =
-        DbLocation::Custom(cache_storage_path().expect("failed to get cache stogate path"));
+    let db_location = DbLocation::Custom(cache_storage_path()?);
     let cache_writer = ImageCacheWriter::as_session(ImageCacheConfig {
         cache_config: CacheConfig {
             db_location,